+++ RxCocoa/iOS/UIAlertView+Rx.swift
@@ -14,10 +14,18 @@ import RxSwift
 
 extension UIAlertView {
     
+    /**
+    Reactive wrapper for `delegate`.
+    
+    For more information take a look at `DelegateProxyType` protocol documentation.
+    */
     public var rx_delegate: DelegateProxy {
         return proxyForObject(self) as RxAlertViewDelegateProxy
     }
     
+    /**
+    Reactive wrapper for `delegate` message.
+    */
     public var rx_clickedButtonAtIndex: ControlEvent<Int> {
         let source = rx_delegate.observe("alertView:clickedButtonAtIndex:")
             .map { a in
@@ -27,6 +35,9 @@ extension UIAlertView {
         return ControlEvent(source: source)
     }
     
+    /**
+    Reactive wrapper for `delegate` message.
+    */
     public var rx_willDismissWithButtonIndex: ControlEvent<Int> {
         let source = rx_delegate.observe("alertView:willDismissWithButtonIndex:")
             .map { a in
@@ -36,6 +47,9 @@ extension UIAlertView {
         return ControlEvent(source: source)
     }
     
+    /**
+    Reactive wrapper for `delegate` message.
+    */
     public var rx_didDismissWithButtonIndex: ControlEvent<Int> {
         let source = rx_delegate.observe("alertView:didDismissWithButtonIndex:")
             .map { a in