+++ RxCocoa/iOS/UIGestureRecognizer+Rx.swift
@@ -51,6 +51,9 @@ class GestureTarget: RxTarget {
 
 extension UIGestureRecognizer {
     
+    /**
+    Reactive wrapper for gesture recognizer events.
+    */
     public var rx_event: ControlEvent<UIGestureRecognizer> {
         let source: Observable<UIGestureRecognizer> = AnonymousObservable { [weak self] observer in
             MainScheduler.ensureExecutingOnScheduler()