+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
 * Deprecates `catchErrorToResult`, the preferred way is to use Swift `do/try/catch` mechanism.
 * Deprecates `RxResult`, the preferred way is to use Swift `do/try/catch` mechanism.
 * Deprecates `sendNext` on `Variable` in favor of just using `value` setter.
+* Renames `rx_searchText` to `rx_text` on `UISearchBar+Rx`.
+* Changes parameter type for `rx_imageAnimated` to be transitionType (kCATransitionFade, kCATransitionMoveIn, ...).
 
 ## [2.0.0-alpha.1](https://github.com/ReactiveX/RxSwift/pull/50) (WIP)